✨ Convert instruction name from IDL into snake case
